Allow providing existing connection to MysqlStorage
I quickly made this change through the Github editor. It would be great if we were allowed to pass an existing connection object into the MysqlStorage object.
My use case is that the storage is plugged into an existing system where connections are provided on demand.

Is there a specific reason you want a connection string explicitly?